# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/subjects/subject": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "summary": "Create Subject",
        "operationId": "create_subject_api_subjects_subject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/subject/{id}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "summary": "Read Subject",
        "operationId": "read_subject_api_subjects_subject__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Subjects"
        ],
        "summary": "Update Subject",
        "operationId": "update_subject_api_subjects_subject__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subjects"
        ],
        "summary": "Delete Subject",
        "operationId": "delete_subject_api_subjects_subject__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deck_configs/deck_config": {
      "post": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Create Deck Config",
        "operationId": "create_deck_config_api_deck_configs_deck_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deck_configs/deck_config/{deck_config_id}": {
      "get": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Read Deck Config",
        "operationId": "read_deck_config_api_deck_configs_deck_config__deck_config_id__get",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Update Deck Config",
        "operationId": "update_deck_config_api_deck_configs_deck_config__deck_config_id__patch",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Delete Deck Config",
        "operationId": "delete_deck_config_api_deck_configs_deck_config__deck_config_id__delete",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/page/subjects/subject/{id}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "summary": "Read Subject",
        "operationId": "read_subject_page_subjects_subject__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/page/deck_configs/deck_config/{id}": {
      "get": {
        "tags": [
          "Deck Configurations"
        ],
        "summary": "Read Deck Config",
        "operationId": "read_deck_config_page_deck_configs_deck_config__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeckConfigCreate": {
        "properties": {
          "deck_id": {
            "type": "string",
            "format": "uuid",
            "title": "Deck Id"
          },
          "prompt_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Answer Pool Counts"
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "prompt_fields",
          "answer_fields",
          "prompt_pool",
          "prompt_pool_counts",
          "answer_pool",
          "answer_pool_counts"
        ],
        "title": "DeckConfigCreate"
      },
      "DeckConfigRead": {
        "properties": {
          "deck_id": {
            "type": "string",
            "format": "uuid",
            "title": "Deck Id"
          },
          "prompt_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Answer Pool Counts"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "prompt_fields",
          "answer_fields",
          "prompt_pool",
          "prompt_pool_counts",
          "answer_pool",
          "answer_pool_counts",
          "id"
        ],
        "title": "DeckConfigRead"
      },
      "DeckConfigUpdate": {
        "properties": {
          "deck_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deck Id"
          },
          "prompt_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Pool Counts"
          }
        },
        "type": "object",
        "title": "DeckConfigUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SubjectCreate"
      },
      "SubjectRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "deck_count": {
            "type": "integer",
            "title": "Deck Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "deck_count"
        ],
        "title": "SubjectRead"
      },
      "SubjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "SubjectUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}